Update the manifest for the release tinker_board_3-android14-1.0.7
The following 2 repositories have large files which can not be pushed
to GitHub and are not needed for Tinker Board 3/3S.

rockchip-android-external-camera_engine_rkaiq
rockchip-android-hardware-rk29-camera360

We need to remove the dependency in the following 2 repositories.

rockchip-android-hardware-rk29-camera_v3
rockchip-android-device-rockchip-rksdk

However, since tinker_board_3-android14-1.0.7 is already created, we
need to update the manifest to use the new revision for the above 2
repositories.

Change-Id: I5e87239419fe6a6454b08fc900b322586b9cc3d0